requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
altgraph==0.17.3
amqp==5.2.0
anyio==4.3.0
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
beautifulsoup4==4.12.3
behave==1.2.6
billiard==4.2.0
bs4==0.0.1
cachetools==5.3.3
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.2.0
crispy-bootstrap4==2022.1
cssselect==1.2.0
cycler==0.12.1
decorator==4.4.2
discord==2.3.2
discord.py==2.3.2
Django==5.0
django-adminlte3==0.1.6
django-appconf==1.0.6
django-autocomplete-light==3.9.7
django-crispy-forms==2.0
django-datatables-view==1.20.0
django-rest-framework==0.1.0
django-select2==8.1.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
execnet==2.0.2
fasteners==0.19
filelock==3.13.3
flatbuffers==24.3.25
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.3.1
gast==0.5.4
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.62.2
gunicorn==21.2.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.4
httpx==0.26.0
huggingface-hub==0.23.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
iniconfig==2.0.0
intel-openmp==2021.4.0
Jinja2==3.1.3
joblib==1.3.2
keras==2.15.0
kiwisolver==1.4.5
kombu==5.3.4
langid==1.1.6
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.1
llvmlite==0.42.0
loguru==0.7.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.2.0
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mysql-connector==2.2.9
Naked==0.1.32
namex==0.0.8
networkx==3.3
numba==0.59.1
numpy==1.25.2
oauthlib==3.2.2
opencv-python==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.11.0
outcome==1.3.0.post0
packaging==24.0
pandas==2.0.3
parameterized==0.9.0
parse==1.20.1
parse-type==0.6.2
pdbp==1.5.0
pefile==2023.2.7
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
pooch==1.8.1
proglog==0.1.10
prompt-toolkit==3.0.41
protobuf==4.25.3
psutil==5.9.8
psycopg2==2.9.6
py==1.11.0
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pycparser==2.21
pycryptodome==3.20.0
PyGetWindow==0.0.9
Pygments==2.17.2
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.6
PyMsgBox==1.0.9
pynose==1.5.0
pyotp==2.9.0
pyparsing==3.1.2
pyperclip==1.8.2
pyreadline3==3.4.1
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytest==8.1.1
pytest-html==2.0.1
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==14.0
pytest-xdist==3.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.8
pytweening==1.2.0
pytz==2023.3
pywin32-ctypes==0.2.2
PyYAML==6.0.1
regex==2024.4.28
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
sacremoses==0.1.1
sbvirtualdisplay==1.3.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
selenium==4.18.1
seleniumbase==4.24.12
shellescape==3.8.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
sqlparse==0.4.4
sympy==1.12
tabcompleter==1.3.0
tbb==2021.12.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.4.0
tokenizers==0.12.1
tomli==2.0.1
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.2
transformers==4.18.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2023.3
ultralytics==8.2.11

vine==5.1.0
wcwidth==0.2.12
Werkzeug==3.0.2
whitenoise==6.6.0
win32-setctime==1.1.0
wrapt==1.14.1
wsproto==1.2.0
yarl==1.9.4
